backend/requirements.txt
annotated-types==0.7.0
anthropic==0.67.0
anyio==3.7.1
APScheduler==3.10.4
av==11.0.0
beautifulsoup4==4.12.2
blinker==1.9.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
coloredlogs==15.0.1
cryptography==45.0.6
ctranslate2==4.6.0
distro==1.9.0
dnspython==2.7.0
dotenv==0.9.9
exceptiongroup==1.3.0
faster-whisper==0.10.0
ffmpeg-python==0.2.0
filelock==3.18.0
Flask==2.3.3
Flask-Cors==4.0.0
flatbuffers==25.2.10
fsspec==2025.7.0
future==1.0.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.4
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
mpmath==1.3.0
numpy==2.0.2
onnxruntime==1.19.2
openai==1.101.0
packaging==25.0
pandoc==2.4
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
plumbum==1.9.0
ply==3.11
protobuf==6.31.1
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pymongo==4.6.0
PyPDF2==3.0.1
pypdfium2==4.30.0
pytesseract==0.3.13
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.2
requests==2.31.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sympy==1.14.0
tokenizers==0.15.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uuid==1.30
Werkzeug==3.1.3
zipp==3.23.0
